+++ src/pages/_app.tsx
@@ -1,11 +1,6 @@
-import "../styles/globals.css";
+import "@/styles/globals.css";
 import "@interchain-ui/react/styles";
 
-import { wallets as cosmostationWallets } from "@cosmos-kit/cosmostation-extension";
-import { wallets as keplrWallets } from "@cosmos-kit/keplr-extension";
-import { wallets as leapWallets } from "@cosmos-kit/leap-extension";
-import { wallets as metamaskWallets } from "@cosmos-kit/leap-metamask-cosmos-snap";
-import { wallets as okxWallets } from "@cosmos-kit/okxwallet";
 import { ChainProvider } from "@cosmos-kit/react";
 import * as RadixToast from "@radix-ui/react-toast";
 import { QueryClientProvider } from "@tanstack/react-query";
@@ -14,34 +9,16 @@ import { AppProps } from "next/app";
 import Head from "next/head";
 import { ComponentProps } from "react";
 import { WagmiConfig } from "wagmi";
-import { configureChains, createConfig } from "wagmi";
-import { MetaMaskConnector } from "wagmi/connectors/metaMask";
-import { publicProvider } from "wagmi/providers/public";
 
 import { getAssetLists, getChains } from "@/chains";
 import { BuildInfo } from "@/components/BuildInfo";
 import MainLayout from "@/components/MainLayout";
-import { EVM_CHAINS } from "@/constants/constants";
 import { AssetsProvider } from "@/context/assets";
 import { ToastProvider } from "@/context/toast";
+import { wallets } from "@/lib/cosmos-kit";
+import { queryClient } from "@/lib/react-query";
+import { wagmiConfig } from "@/lib/wagmi";
 import { SkipProvider } from "@/solve";
-import { queryClient } from "@/utils/query";
-import { OKXConnector } from "@/wallets/OKXConnector";
-
-const { publicClient, chains: evmChains } = configureChains(EVM_CHAINS, [
-  publicProvider(),
-]);
-
-const wagmiConfig = createConfig({
-  autoConnect: true,
-  connectors: [
-    new MetaMaskConnector({
-      chains: evmChains,
-    }),
-    OKXConnector,
-  ],
-  publicClient,
-});
 
 type ChainProviderProps = ComponentProps<typeof ChainProvider>;
 
@@ -49,14 +26,6 @@ const assets = getAssetLists() as ChainProviderProps["assetLists"];
 const chains = getChains() as ChainProviderProps["chains"];
 
 export default function App({ Component, pageProps }: AppProps) {
-  const wallets = [
-    ...keplrWallets,
-    ...cosmostationWallets,
-    ...leapWallets,
-    ...metamaskWallets,
-    ...okxWallets,
-  ];
-
   return (
     <>
       <Head>
